requirements.txt
pydub
simpleaudio
ffmpeg-python
pyopenssl>=23.0.0
websocket-client>=1.5.1